Move the lyric row state into the inherit class.
