+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/EditLyricRow.cs
@@ -3,6 +3,9 @@
 
 #nullable disable
 
+using System;
+using osu.Framework.Allocation;
+using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Edit;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Info;
@@ -10,14 +13,38 @@ using osu.Game.Rulesets.Karaoke.Objects;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows
 {
-    public class EditLyricRow : LyricEditorRow
+    [Cached(typeof(IEditLyricRowState))]
+    public class EditLyricRow : LyricEditorRow, IEditLyricRowState
     {
         private const int min_height = 75;
+
+        private readonly IBindable<LyricEditorMode> bindableMode = new Bindable<LyricEditorMode>();
+        private readonly IBindable<int> bindableLyricPropertyWritableVersion;
+
+        public event Action<LyricEditorMode> WritableVersionChanged;
 
         public EditLyricRow(Lyric lyric)
             : base(lyric)
         {
             AutoSizeAxes = Axes.Y;
+
+            bindableLyricPropertyWritableVersion = lyric.LyricPropertyWritableVersion.GetBoundCopy();
+        }
+
+        [BackgroundDependencyLoader]
+        private void load(ILyricEditorState state)
+        {
+            bindableMode.BindTo(state.BindableMode);
+
+            bindableMode.BindValueChanged(x =>
+            {
+                WritableVersionChanged?.Invoke(bindableMode.Value);
+            });
+
+            bindableLyricPropertyWritableVersion.BindValueChanged(_ =>
+            {
+                WritableVersionChanged?.Invoke(bindableMode.Value);
+            });
         }
 
         protected override Drawable CreateLyricInfo(Lyric lyric)
